change gsed to sed & fixing several issues
In the current version of Bitrise, this step can no longer be used in a workflow out of the box. Bitrise does not recognize the gsed (gnu-sed) command as a command, resulting in errors. Therefore gsed was replaced with sed and in the course of this the script was also simplified and small errors were corrected.